Finalized pictures controller Moved DTO creation into its constructor

+++ sephora-backend/perfume_luxury_web_api/Controllers/PicturesController.cs
@@ -1,34 +1,50 @@
 namespace perfume_luxury_web_api.Controllers;
 
-[ApiController, Route("api/[controller]")]
+[ApiController, Route("assets/img")]
 public class PicturesController(
-    IPictureService pictureService
+    IPictureService pictureService,
+    IHostEnvironment env
 ) : ControllerBase
 {
     [HttpPost]
     public async Task<IActionResult> SaveImage(IFormFile file)
     {
         string picName = await pictureService.SaveImage(file);
-        return Ok(pictureService.CreatePicDto(picName));
+        return Ok(new PictureDto(picName, env.IsDevelopment()));
     }
 
-    // [HttpGet("{name}")]
-    // public IActionResult GetImage([FromRoute] string name)
-    // {
-    //     if (!pictureService.FileExists(name))
-    //         return NotFound(new
-    //         {
-    //             Status = "404 Not Found",
-    //             Error = "Image not found"
-    //         });
-    //     
-    //     return File(pictureService.GetFile(name), "image/webp");
-    // }
-
+    [HttpGet("{name}")]
+    public IActionResult GetImage(
+        [FromRoute] string name, 
+        [FromQuery] string size = "original"
+            )
+    {
+        if (!pictureService.SizeExists(size)) 
+            return NotFound(new
+            {
+                Status = "404 Not Found",
+                Error = "Invalid size"
+            });
+        if (!pictureService.FileExists(name))
+            return NotFound(new
+            {
+                Status = "404 Not Found",
+                Error = "Image not found"
+            });
+        
+        return File(pictureService.GetFile(name, size), "image/webp");
+    }
+    
     // may be FromBody|FromQuery
     [HttpDelete("{name}")]
     public IActionResult DeleteImage([FromRoute] string name)
     {
+        if (!pictureService.FileExists(name))
+            return NotFound(new
+            {
+                Status = "404 Not Found",
+                Error = "Image not found"
+            });
         try
         {
             pictureService.DeleteFile(name);
